+++ assets.go
@@ -16,6 +16,9 @@ var appCSS []byte
 //go:embed web/app.js
 var appJS []byte
 
+//go:embed web/gba-text.js
+var gbaTextJS []byte
+
 //go:embed web/menu-themes.js
 var menuThemesJS []byte