+++ player/build.sh
@@ -11,7 +11,7 @@ OBJCOPY=${OBJCOPY:-llvm-objcopy}
 rm -rf "$BUILD"
 mkdir -p "$BUILD"
 
-COMMON="--target=arm-none-eabi -mcpu=arm7tdmi -marm -ffreestanding -fno-builtin -fno-stack-protector -Os -Wall -Wextra"
+COMMON="--target=arm-none-eabi -mcpu=arm7tdmi -marm -ffreestanding -fno-builtin -fno-stack-protector -Os -fno-inline-functions -Wall -Wextra"
 $CLANG $COMMON -c "$PLAYER/startup.S" -o "$BUILD/startup.o"
 $CLANG $COMMON -c "$PLAYER/metadata.S" -o "$BUILD/metadata.o"
 $CLANG $COMMON -c "$PLAYER/main.c" -o "$BUILD/main.o"